requirements.txt
#FastAPI & Server
fastapi
uvicorn[standard]
python-dotenv
python-multipart

#AI & Machine Learning Core
tensorflow
torch
torchvision
transformers
scikit-learn
pandas
numpy
opencv-python-headless

#LangChain & OpenAI
langchain
langchain-openai
langchain-chroma
langchain-community
python-dotenv
tiktoken
duckduckgo-search
chromadb
beautifulsoup4